Header aligns properly after reordering columns
Before, in tables with a fixed column, you could cause the column headers to get out of alignment with the rest of the table by scrolling a bit then reordering the columns.

This change makes it so that sidescrolling gets triggered when the columns change. This also removes references to scrolling from header rows (as this should be handled in the header block).